+++ src/lib/ColorPicker.svelte
@@ -3,24 +3,41 @@
   import ColorArea from './ColorArea.svelte'
   import HueSlider from './HueSlider.svelte'
   import AlphaSlider from './AlphaSlider.svelte'
+  import { shouldShowAbove } from '$lib'
 
   export let color: Color
   export let isOpen = false
   export let showAlphaSlider = false
+  /** Element used to figure out position (probably an input element) */
+  export let positioningContextElement: HTMLElement
+  let pickerEl: HTMLDivElement | undefined
 
   export let onInput = () => {
     /* noop */
   }
   $: color, onInput()
+
+  let showAbove = false
+  $: if (pickerEl && positioningContextElement) {
+    showAbove = shouldShowAbove(pickerEl, positioningContextElement)
+  }
 </script>
 
-<div class="color-picker" class:hidden={!isOpen} on:touchstart|preventDefault>
-  <ColorArea bind:color {onInput} />
-  <HueSlider bind:color {onInput} />
-  {#if showAlphaSlider}
-    <AlphaSlider bind:color {onInput} />
-  {/if}
-</div>
+{#if isOpen}
+  <div
+    bind:this={pickerEl}
+    class="color-picker"
+    class:above={showAbove}
+    class:hidden={!isOpen}
+    on:touchstart|preventDefault
+  >
+    <ColorArea bind:color {onInput} />
+    <HueSlider bind:color {onInput} />
+    {#if showAlphaSlider}
+      <AlphaSlider bind:color {onInput} />
+    {/if}
+  </div>
+{/if}
 
 <style lang="sass">
   .color-picker
@@ -39,6 +56,9 @@
     //   width: 170px
     //   height: 170px
     z-index: 50
+    &.above
+      top: auto
+      bottom: 100%
     &.hidden
       display: none
     :global(.slider)