+++ src/lib/index.ts
@@ -7,3 +7,13 @@ export { default as ColorArea } from './ColorArea.svelte'
 export { default as Slider } from './Slider.svelte'
 
 export { Color } from './color'
+
+/** Determines if `popupElement` should be shown above `positioningContextElement` */
+export function shouldShowAbove(popupElement: HTMLElement, positioningContextElement: HTMLElement) {
+  const inputBounds = positioningContextElement.getBoundingClientRect()
+  const spaceAbove = inputBounds.top
+  const spaceBelow = window.innerHeight - (inputBounds.top + inputBounds.height)
+  const enoughSpaceAbove = spaceAbove > popupElement.clientHeight
+  const enoughSpaceBelow = spaceBelow > popupElement.clientHeight
+  return !enoughSpaceBelow && enoughSpaceAbove
+}