+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Add auto positioning
 - Open color picker if `ColorInput` is already focused when mounting
 
 ## v1.2.0 - 2023 Apr 17